% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_deviations.R
\name{build_deviations}
\alias{build_deviations}
\title{Retrieve data from NOAA API and calculate deviations between observed and
predicted water levels}
\usage{
build_deviations(
  .station,
  .first_yr,
  .last_yr,
  .datum = c("MSL", "MTL", "MHW", "MHHW", "MLLW", "MLW", "NAVD", "STND"),
  .units = c("metric", "english"),
  .timefmt = c("gmt", "lst", "lst_ldt"),
  .tz = "UTC"
)
}
\arguments{
\item{.station}{Numeric or Character. The code that uniquely identifies a
NOAA water level station.}

\item{.first_yr}{Integer. First year from which to retrieve data.}

\item{.last_yr}{Integer. Last year from which to retrieve data.}

\item{.datum}{Character string, from a limited list of options. Specifies the
vertical datum used for the returned data. "STND" is a
station-specific datum anticipated to be below all observed or
predicted water levels, so al lwater levels wil lbe positive numbers.
"NAVD" is the National Vertical Datum, useful for relating tides to
ground elevations from survey or LIDAR.  The other choices are
standard tidal datums.}

\item{.units}{"metric" or "english".  The default is "metric".}

\item{.timefmt}{'gmt', 'lst', or 'lst_ldt'.  How would you like the time
coordinate defined?  Choices are for greenwich mean time (also known as
UTC, although that abbreviation does not work here), local standard
time (with no adjustments for daylight savings time), or local clock
time, which can be confusing, since the data returned is a string (not
an internal \code{POSIXct} or \code{POSIXlt} object).  Interpretting the data
correctly requires care specifying the timezone.}

\item{.tz}{An R timezone specification. Used only for converting text-based
times deliverd from the API to POSIXct objects. Defaults to 'UTC', for
consistency with the default \code{.timefmt = 'gmt'}. For local standard
time, as specified via \code{.timefmt = 'lst'}, the easiest way is to
specify an offset from UTC. For example, use  'Etc/GMT+5' for the U.S.
East Coast.  For \code{.timefmt = 'lst_ldt'} you will need to figure out
how to define your local clock time as a time zone. For the U.S East
Coast, \code{.tz = "America/New_York"} works. See \code{?timezone}
for more information.}
}
\value{
A dataframe with columns "datetime", "water_level" and "deviation".
The values in "deviation" are the difference in observed and
predicted tides. Positive values mean observed water level was higher
than predicted.
}
\description{
Retrieve data from NOAA API and calculate deviations between observed and
predicted water levels
}
