% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{call_api}
\alias{call_api}
\title{Send One call to the NOAA API to retrieve hourly water level data}
\usage{
call_api(
  .station,
  .start,
  .stop,
  .which = c("observed", "predicted"),
  .datum = c("MSL", "MTL", "MHW", "MHHW", "MLLW", "MLW", "NAVD", "STND"),
  .units = c("metric", "english"),
  .timefmt = c("gmt", "lst", "lst_ldt"),
  .tz = "UTC"
)
}
\arguments{
\item{.station}{Numeric or Character. The code that uniquely identifies a
NOAA water level station.}

\item{.start}{An object that inherits from R's \code{Date} class or a string of
either "yyyymmdd" or "mm/dd/yyyy" format.  The starting date for the
period of data retrieval.}

\item{.stop}{An object that inherits from R's \code{Date} class or a string of
either "yyyymmdd" or "mm/dd/yyyy" format.  The ending date for the
period of data retrieval.  The NOAA API restricts data retrieval of
hourly data to no more than one year.  This function will throw an
error without sending a request to the NOAA API if your request does
not comply with that requirement.}

\item{.which}{'observed' or 'predicted'}

\item{.datum}{Character string, from a limited list of options. Specifies the
vertical datum used for the returned data. "STND" is a
station-specific datum anticipated to be below all observed or
predicted water levels, so al lwater levels wil lbe positive numbers.
"NAVD" is the National Vertical Datum, useful for relating tides to
ground elevations from survey or LIDAR.  The other choices are
standard tidal datums.}

\item{.units}{"metric" or "english".  The default is "metric".}

\item{.timefmt}{'gmt', 'lst', or 'lst_ldt'.  How would you like the time
coordinate defined?  Choices are for greenwich mean time (also known as
UTC, although that abbreviation does not work here), local standard
time (with no adjustments for daylight savings time), or local clock
time, which can be confusing, since the data returned is a string (not
an internal \code{POSIXct} or \code{POSIXlt} object).  Interpretting the data
correctly requires care specifying the timezone.}

\item{.tz}{An R timezone specification. Used only for converting text-based
times deliverd from the API to POSIXct objects. Defaults to 'UTC', for
consistency with the default \code{.timefmt = 'gmt'}. For local standard
time, as specified via \code{.timefmt = 'lst'}, the easiest way is to
specify an offset from UTC. For example, use  'Etc/GMT+5' for the U.S.
East Coast.  For \code{.timefmt = 'lst_ldt'} you will need to figure out
how to define your local clock time as a time zone. For the U.S East
COast, \code{.tz = "America/New_York"} works. See \code{?timezone}
for more information.}
}
\value{
A dataframe with columns "datetime" and "water_level"
}
\description{
\code{call_api()} places a single call to the NOAA API to retrieve either
hourly water level observations or predictions.
}
\examples{
Providence_station <- 8454000
begin_date <- '20200806'
end_date   <- '20200810'
call_api_observed(Providence_station, begin_date, end_date)

}
\concept{NOAA water level data access functions}
