% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_timezone}
\alias{.is_timezone}
\title{Check if a timezone specification is likely to work}
\usage{
.is_timezone(.timezone)
}
\arguments{
\item{timezone}{}
}
\value{
Boolean.
}
\description{
Compares a candidate timezone string to the list of recognized time zones in
the \code{OlsonNames()} list. available in base R.  While this is not necessarily
a complete list of valid time zones (which issystem-specific), it's a good
portlable starting point.
}
\examples{
.is_timezone(c('Etc/GMT+5', 'EST', 'America/New_York'))
}
