% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_whole}
\alias{.is_whole}
\title{Check if a value is (close to) an Integer}
\usage{
.is_whole(.x, .tol = .Machine$double.eps^0.5)
}
\arguments{
\item{.x}{Value or numerical vector to check}

\item{.tol}{The tolerance.  How close to an integer do you want to count as an
integer?  Default value is related to precision of floats, and
thus depends on the computer and specific R installation.  The
default may be far too precise for many use cases, where you want
to accept numbers somewhat close to an integer value.}
}
\value{
Boolean (\code{TRUE} or \code{FALSE}) vector, with length matching the
length of \code{.x}.
}
\description{
Function is based on the code provided as an example for the function
\code{base::is.integer()} (which does NOT do what you might expect).
}
\examples{
.is_whole(4)
.is_whole(4.2)
.is.wholer(4.2, .tol = 0.25)
}
