% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.lst_2_df}
\alias{.lst_2_df}
\title{Utility function to extract data from a list of lists to a data frame}
\usage{
.lst_2_df(lst_o_lsts, tz = "UTC")
}
\arguments{
\item{lst_o_lsts}{Source data, as a list of lists, as created by.}

\item{tz}{Designation of timezone, to correctly convert from string to R
dates and times.  Correct value depends on the parameter sent to the
NOAA API when requesting the data. Defaults to 'Etc/GMT+4',
which is east coast US standard time.}
}
\value{
a tibble
}
\description{
Applying \code{httr::content()} to the (json) data returned by \code{httr::GET()}
request sent to the NOAA API generates a nested list of lists.  Each sub-list
has four (five) elements, including t = the time , and v = the value (water
depth).  This function takes the list of lists and converts it into a
dataframe with two data columns datetime (a 'POSIXct' value), and
water_depth.
}
