% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodcast_ar.R
\name{.sim_once}
\alias{.sim_once}
\title{Generate a Single Simulation of Tidal Deviations}
\usage{
.sim_once(
  .data,
  .dt,
  .predicted,
  .model,
  .slr,
  .fldlvl,
  .wl_update = 0,
  .len = NULL
)
}
\arguments{
\item{.data}{A source data frame. can be NULL if .dt and .wl are defined
in the enclosing environment.}

\item{.dt}{Dates and times.  A data column in .data or a data vector
defined in the enclosing environment that contains date-time
information.  Must inherit from POSIXct.}

\item{.predicted}{The (astronomical) tidal predictions. A numerical data
column in .data or a numerical vector defined in the enclosing
environment.}

\item{.model}{a time series model with a suitable \code{simulate()} method,
usually an ARIMA model model of \code{class = c("ARIMA")}.}

\item{.slr}{Sea level rise estimate. A numeric vector containing one or
more values. What sea level rise scenarios should the analysis
be based on?  Selection of appropriate sea level rise
estimates should be based on your time horizon and risk
tolerance. NOAA forecasts of SLR for 2100 vary from 0.3
(almost certain) to 2.5 (highly unlikely) meters.}

\item{.fldlvl}{Flood level.  A single numeric value, that defines what
constitutes a flood event. Must be specified in
\code{floodcast_tub()}. In \code{floodcast_tub_lookup()}, will default
to the "Highest Astronomical Tide" elevation, or HAT for the
selected tide station.}

\item{.wl_update}{A single numeric value that adjusts for any changes in
sea level from the tidal epoch to the present day.}

\item{.len}{Desired length of simulated output. IF \code{NULL}, defaults to the
length of the input deviations.}
}
\value{
numerical vector of length equal to input vector
}
\description{
Simulates a single random time series, based on an ARIMA model.  Can accept
regression parameters as well, usually Year and an annual (season) fourier
decomposition. terms
}
\details{
For now, both .nonzero and .seasonal are needed to correctly interpret the \code{.coefs}
parameter. That may change as we revise this function, as the
}
\examples{

}
\seealso{
Other Flood frequency analysis functions: 
\code{\link{floodcast_tub}()},
\code{\link{floodfreq}()},
\code{\link{floodgraph}()},
\code{\link{floodmean}()}
}
\concept{Flood frequency analysis functions}
