% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodcast_ar.R
\name{floodcast_ar}
\alias{floodcast_ar}
\title{Forecasting future coastal flood frequency based on time series simulation}
\usage{
floodcast_ar(.data, .dt, .wl, .slr, .fldlvl, .wl_offset = 0)
}
\arguments{
\item{.data}{A source data frame. can be NULL if .dt and .wl are defined
in the enclosing environment.}

\item{.dt}{Dates and times.  A data column in .data or a data vector
defined in the enclosing environment that contains date-time
information.  Must inherit from POSIXct.}

\item{.wl}{Water level.  A numerical data column in .data or a numerical
vector defined in the enclosing environment.}

\item{.slr}{Sea level rise estimate. A numeric vector containing one or
more values. What sea level rise scenarios should the analysis
be based on?  Selection of appropriate sea level rise
estimates should be based on your time horizon and risk
tolerance. NOAA forecasts of SLR for 2100 vary from 0.3
(almost certain) to 2.5 (highly unlikely) meters.}

\item{.fldlvl}{Flood level.  A single numeric value, that defines what
constitutes a flood event. Must be specified in
\code{floodcast_tub()}. In \code{floodcast_tub_lookup()}, will default
to the "Highest Astronomical Tide" elevation, or HAT for the
selected tide station.}
}
\description{
Bootstrap estimates of future annual flood frequency by adding simulated
deviations to predicted tides.
}
\details{
The function simulates "random" deviations from tidal predictions in a
manner that creates deviations with statistical properties similar to those
of the real deviations observed in the past. We bootstrap the simulations to
produce estimates and standard errors of future annual flood frequencies.

The simulations are based on an ARMA process derived from historic data. An
ARMA process is a specific flavor of time series model. (The acronym
reflects the fact that the model includes both "autoregressive" and "moving
average" model components.)

By definition, the observed deviations during the tidal epoch have
mean zero (or very close to zero). Deviations not during the tidal epoch may
have either positive or negative mean.

Ideally, they would also be stationary,
but in practice they often are not.  As we explain elsewhere, the deviations
often have a small positive slope, as the tidal predictions did not take into
account sea level rise. The effect is small, and we chose to overlook it for
modeling purposes.
}
