% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodcast_arima.R
\name{floodcast_arima}
\alias{floodcast_arima}
\title{Forecasting future coastal flood frequency based on time series simulation}
\usage{
floodcast_arima(x, y, z)
}
\arguments{
\item{z}{}
}
\value{

}
\description{
Forecasting future coastal flood frequency based on time series simulation
}
