% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodcast_tub.R
\name{floodcast_tub}
\alias{floodcast_tub}
\title{Forecasting future coastal flood frequency using a "Bathtub" model}
\usage{
floodcast_tub(x, y, z)
}
\arguments{
\item{z}{}
}
\value{

}
\description{
Forecasting future coastal flood frequency using a "Bathtub" model
}
