% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodfreq.R
\name{floodfreq}
\alias{floodfreq}
\title{Count Number of Floods and Estimate Daily Risk and Annual Flood Frequency}
\usage{
floodfreq(.data, .dt, .wl, .fldlvl)
}
\arguments{
\item{.data}{A source data frame. can be NULL if .dt and .wl are defined
in the enclosing environment.}

\item{.dt}{Dates and times.  A data column in .data or a data vector
defined in the enclosing environment that contains date-time
information.  Must inherit from POSIXct.}

\item{.wl}{Water level.  A numerical data column in .data or a numerical
vector defined in the enclosing environment.}

\item{.fldlvl}{The water level that defines a flood event. Needs to be
expressed in the same units and the same datum as the water
level data.)}
}
\value{
A named numeric vector consisting of the values \code{days}, \code{flood_days},
\code{daily_p_flood}, and \code{floods_p_yr}.
}
\description{
This function splits a data set by day and counts the number of days in
the data set where the water level (\code{.wl}) exceeded a specified flood
elevation (\code{.fldlvl}) at least once.  It then calculates the proportion of
days with flooding (as an estimate of daily flood risk) and the expected
number of days with flood events per year (by scaling by the daily risk
by 365.25 days per year).
}
\details{
There is no effort to account for partial data within days. This function is
principally intended for exploratory data review or as a building block for
more complex analyses. The function, \code{floodmeans()} provides closely
related statistics, averaged on a year-by-year basis, and offering
estimates of year to year variability. \code{floodcounts()} provides the raw
annual counts of the number of days with flooding.
}
\examples{
old <- prov_tides[prov_tides$Year >= 1940 & prov_tides$Year < 1950,]
new <- prov_tides[prov_tides$Year >= 2010 & prov_tides$Year < 2020,]
floodcount(old, DateTime, MLLW, 1.987)
floodcount(new, DateTime, MLLW, 1.987)

}
\seealso{
Other Flood frequency analysis functions: 
\code{\link{.sim_once}()},
\code{\link{floodcast_tub}()},
\code{\link{floodgraph}()},
\code{\link{floodmean}()}
}
\concept{Flood frequency analysis functions}
