% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodmean.R
\name{floodmean}
\alias{floodmean}
\title{Calculate Mean Days per Year with Flood Events From Historical Data}
\usage{
floodmean(.data, .dt, .wl, .fldlvl)
}
\arguments{
\item{.data}{A source data frame. can be NULL if .dt and .wl are defined
in the enclosing environment.}

\item{.dt}{Dates and times.  A data column in .data or a data vector
defined in the enclosing environment that contains date-time
information.  Must inherit from POSIXct.}

\item{.wl}{Water level.  A numerical data column in .data or a numerical
vector defined in the enclosing environment.}

\item{.fldlvl}{The water level that defines a flood event. Needs to be
expressed in the same units and the same datum as the water
level data.)}
}
\value{
A 3 (row) x 2 (column) array containing \code{Mean} and \code{Var} for
\itemize{
\item \verb{Daily Probability} of flooding,
\item Annual \verb{Flood Days} observed, and
\item Annual \verb{Scaled Flood Days}, which are scaled to account for days each year
with no data.
}
}
\description{
This function takes high frequency data on tidal elevations (typically
hourly or more frequently) and counts the number of days each
year where the water level (\code{.wl}) exceeded a specified flood
elevation (\code{.fldlvl}) at least once.  It calculates the mean annual
number of days with observed flooding, the daily probability of flooding,
and a scaled estimate of days with flooding, scaled to account for days with
no water level data.
}
\examples{
old <- prov_tides[prov_tides$Year >= 1940 & prov_tides$Year < 1950,]
new <- prov_tides[prov_tides$Year >= 2010 & prov_tides$Year < 2020,]
floodmean(old, DateTime, MLLW, 1.987)
floodmean(new, DateTime, MLLW, 1.987)

}
\seealso{
Other Flood frequency analysis functions: 
\code{\link{.sim_once}()},
\code{\link{floodcast_tub}()},
\code{\link{floodfreq}()},
\code{\link{floodgraph}()}
}
\concept{Flood frequency analysis functions}
