% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_availability.R
\name{get_availability}
\alias{get_availability}
\title{Retrieve period of operation of a NOAA Tide Station}
\usage{
get_availability(.station)
}
\arguments{
\item{.station}{The code specifying the station of interest.}
}
\value{
A list with two components: \verb{$firstyr} and \verb{$lastyr} for the year
the station was established (not  when its current configuration was
put into place) and pulled from service, respectively. If the station
is still in service, \verb{$lastyr} will return the current year.
}
\description{
Retrieve period of operation of a NOAA Tide Station
}
\examples{
 Providence_station <- 8454000
 get_availability(Providence_station)
}
\seealso{
Other station information access functions: 
\code{\link{get_datums}()},
\code{\link{get_epoch}()},
\code{\link{get_hat}()},
\code{\link{get_sl_trend}()},
\code{\link{get_tz}()}
}
\concept{station information access functions}
