% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datums.R
\name{get_datums}
\alias{get_datums}
\title{Retrieve a named vector of tidal datums}
\usage{
get_datums(.station, .units = "metric")
}
\arguments{
\item{.station}{A number or string specifying the NOAA station for which
to retreive metadata. A list of NOAA stations is available at
https://tidesandcurrents.noaa.gov/stations.html?type=Water+Levels. A
user-friendly map interface is available here:
https://tidesandcurrents.noaa.gov/map/index.html?type=TidePredictions.}

\item{.units}{Should values be returned in metric (default) or english units?}
}
\value{
A named vector of numerical values.  The names are the abbreviations for the
tidal datum, facilitating data access.

The values are as provided by NOAA. The list includes several statistics of
the tidal regime that are not reference elevations, but provide other
information about the local tidal regime.

All elevation datums are referenced to the (arbitrary) Station Datum. Not all
datums available on Station web pages are included.  In particular, HAT, LAT,
max and min tides are not included here. See the related \code{get_hat()} function
for access to HAT.
}
\description{
Retrieve information from NOAA on tidal datums and certain statistics
describing the tidal regime at NOAA tide stations.
}
\details{
Each tidal Station in the NOAA Station network measures (or estimates) water
elevation at regular intervals (usually every six minutes, or ten times an
hour). Raw elevations are referenced to a station-specific reference
elevation, called the Station Datum.  Usually, however, we want elevations
expressed in units that can be interpreted in relation to tidal
charts, topographic maps, surveyor's benchmarks or statistical summaries of
water level, such as the mean height of all high tides.

The values returned by this function provide information on the
relative elevation of different elevation reference frames, and thus
facilitate interconversions.

The return value here includes all values provided by NOAA as part of their
list of datums. The list of available datums may not always remain the same.
Most tidal datums are defined as means over the period of the tidal epoch. A
list of datums with descriptions is available from NOAA here:
https://tidesandcurrents.noaa.gov/datum_options.html. Not all datums defined
there are available via this function.

The primary values available include the following:
\describe{
\item{MHHW}{Mean Higher High Water (mean of the higher high tide each day)}
\item{MHW}{Mean High Water (mean of all high tides)}
\item{DTL}{Mean Diurnal Tide Level (mean of MHHW and MLLW. Sometimes MDTL)}
\item{MTL}{Mean Tidel Level.  Mean of MHW and MLW. (MHW + MLW)/2)}
\item{MSL}{Mean Sea Level.  Mean of hourly tidal observations over the tidal
epoch.}
\item{MLW}{Mean Low Water}
\item{MLLW}{Man Lower Low Water}
\item{GT}{Greater Diurnal Range. (MHHW - MLLW)}
\item{MN}{Mean Range of Tide. (MHW - MLW)}
\item{DHQ}{Mean Diurnal High Water Inequality.  Mean difference between the
height of two daily high tides}
\item{DLQ}{Mean Diurnal Low Water Inequality.  Mean difference between the
height of two daily low tides}
\item{HWI}{Greenwich High Water Interval (hours). The average interval
between the moon's transit over the Greenwich meridian and the
following high water at a location}
\item{LWI}{Greenwich High Water Interval (hours). The average interval
between the moon's transit over the Greenwich meridian and the
following low water at a location}
\item{NAVD88}{North American Vertical Datum of 1988}
}
}
\examples{
# Retreive datums for Portland, Maine
portland_id <- 8418150
get_datums(portland_id)
}
\seealso{
Other station information access functions: 
\code{\link{get_availability}()},
\code{\link{get_epoch}()},
\code{\link{get_hat}()},
\code{\link{get_sl_trend}()},
\code{\link{get_tz}()}
}
\concept{station information access functions}
