% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datums.R
\name{get_epoch}
\alias{get_epoch}
\title{Retrieve the tidal epoch}
\usage{
get_epoch(.station)
}
\arguments{
\item{.station}{}
}
\value{
A named list of two integers, representing the starting and ending
years of the official tidal epoch.
}
\description{
Retrieve the tidal epoch
}
\examples{
portland_id <- 8418150
get_epoch(portland_id)
}
\seealso{
Other Datum Access: 
\code{\link{get_datums}()},
\code{\link{get_hat}()}
}
\concept{Datum Access}
