% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datums.R
\name{get_hat}
\alias{get_hat}
\title{Retrieve the value of the Highest Astronomical Tide (HAT)}
\usage{
get_hat(
  .station,
  .units = "metric",
  .base = c("MSL", "STND", "MHHW", "MHW", "DTL", "MTL", "MLW", "MLLW", "NAVD88")
)
}
\arguments{
\item{.station}{A number or string specifying the NOAA station for which
to retreive metadata. A list of NOAA stations is available at
https://tidesandcurrents.noaa.gov/stations.html?type=Water+Levels. A
user-friendly map interface is available here;
https://tidesandcurrents.noaa.gov/map/index.html?type=TidePredictions.}

\item{.units}{Should values be returned in metric (default) or english units?}

\item{.base}{The base datum against which to reference HAT. Defaults to MSL,
with a warning.}
}
\value{
A decimal value, representing the HAT value at the selected station.
}
\description{
Retrieve the value of the Highest Astronomical Tide (HAT)
}
\examples{
# Retrieve HAT for Portland, Maine
portland_id <- 8418150
get_hat(portland_id)                   # 2.134, with a warning
get_hat(portland_id, .base = 'MSL')    # 2.134, no warning
get_hat(portland_id, .base = 'MLLW')   # 3.64
}
\seealso{
Other station information access functions: 
\code{\link{get_availability}()},
\code{\link{get_datums}()},
\code{\link{get_epoch}()},
\code{\link{get_sl_trend}()},
\code{\link{get_tz}()}
}
\concept{station information access functions}
