% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sl_trend.R
\name{get_sl_trend}
\alias{get_sl_trend}
\title{Retrieve NOAA Long-Term SLR Rate Estimates}
\usage{
get_sl_trend(.station, .affil = c("us", "global", "all"))
}
\arguments{
\item{.station}{Unique identifier for station of interest. See
https://tidesandcurrents.noaa.gov/stations.html?type=Water+Levels}

\item{.affil}{Should matches be restricted to US stations, or global network
stations. Default is 'us'.  If 'all' is specified, then both networks
are searched.}
}
\value{
A list containing specifics about NOAA's long-term SLR estimates.

\describe{
\item{stationId}{Unique NOAA id for accessing the Station.}
\item{stationName}{Name of Station.}
\item{affil}{"US"or "Global"}
\item{latitude}{Position of Station}
\item{longitude}{Position of Station}
\item{trend}{Point estimate of sea level rise trend.}
\item{trendError}{Standard error of trend estimate.}
\item{units}{Apparently always "mm/yr"?}
\item{startDate}{Start of period over which trend was calculated.}
\item{endDate}{End date of period over which trend was caclulated.}
}
}
\description{
This function returns NOAA's long-term estimates of sea level rise
for a specific water level monitoring station, identified by NOAA's unique
identifiers, along with appropriate metadata.
}
\details{
The NOAA API can, in principal return data for more than one station if there
are multiple matches. This function does not yet handle that situation
gracefully. It merely returnsan error.  Users are encouraged to make sure
their search selection of \code{.station} an \code{.affil} return data for only a
single station.
}
\examples{
portland_id <- 8418150
sl_list <- get_sl_trend(portland_id)
names(sl_list)
paste(sl_list$trend, '+/-', sl_list$trendError, 'sl_list$units')

}
\seealso{
Other sea level rate functions: 
\code{\link{slr_change_comp}()},
\code{\link{slr_change}()},
\code{\link{slr_slope}()}

Other station information access functions: 
\code{\link{get_availability}()},
\code{\link{get_datums}()},
\code{\link{get_epoch}()},
\code{\link{get_hat}()},
\code{\link{get_tz}()}
}
\concept{sea level rate functions}
\concept{station information access functions}
