% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_availability.R
\name{get_tz}
\alias{get_tz}
\title{Access station timezone information, either as an offset from UTC or a timezone string}
\usage{
get_tz(.station, .type = c("offset", "string"))
}
\arguments{
\item{.station}{Code for NOAA tide station.}

\item{.type}{Either 'offset' for an integer value offset from UTC, or}
}
\value{
Either an integer  for \code{.type = 'offset'}) or a string
(`.type = 'string')of specifying a time zone offset, of the form
'Etc/GMT+X', where X is an integer. This string form is likely to
only work correctly for the western hemisphere, and has not been
tested elsewhere.  U.S Eastern Standard Time is offset == -5,
or 'Etc/GMT+5'
}
\description{
Consults the NOAA metadata API for your selected station, and queries it
for information on the station's (standard time) timezone offset from UTC.
}
\examples{
Providence_station <- 8454000  # Providence, Rhode Island
get_tz(Providence_station)
get_tz(Providence_station, .type = 'string')
}
\seealso{
Other station information access functions: 
\code{\link{get_availability}()},
\code{\link{get_datums}()},
\code{\link{get_epoch}()},
\code{\link{get_hat}()},
\code{\link{get_sl_trend}()}
}
\concept{station information access functions}
