% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_change.R
\name{slr_change}
\alias{slr_change}
\title{Test whether the rate of sea level rise has changed}
\usage{
slr_change(
  .data,
  .sl,
  .dt,
  .span = 20L,
  .mode = c("year", "time", "count"),
  t_fit = FALSE,
  retain_model = FALSE
)
}
\arguments{
\item{.data}{Source data frame for data.  Use NULL if no data frame is used
and all data is passed from
the enclosing environment.}

\item{.sl}{The data variable (usually found in the data frame) showing sea
level or mean sea level.  Must be a named variable, not an
expression.}

\item{.dt}{Data variable containing corresponding midpoint dates for the
period of averaging used to calculate .sl. Must be a named variable,
not an expression. Midpoint dates for a given month can be approximated
with \verb{as.Date(paste(year, month, 15, sep = '-')}.}

\item{.span}{Span defining the "recent" period for fitting the model.
integer or difftime object.  Interpretation of .span depends on the
value of the by_year. If \code{by_year == TRUE}, \code{.span} must be an
integer, and is interpreted as the number of years included in the
"recent" period for fitting the model.  If \code{by_year == FALSE}, and
\code{.span} is an integer, it is interpreted as the number of records to
include in the 'recent' period.}

\item{.mode}{one of c('year', 'time', 'count') indicating whether the .span
is expressed in years, time coordinates, or number of observations.
If \code{.mode == 'year'}, slopes are scaled to be expressed on a per year
basis.  Scaling requires the time coordinate to inherit from R
\code{Date} or \code{POSIXt} objects.  Rates based on \code{Dates} are converted to
annual values by multiplying by 365.2422 days / year.  Rates based on
\code{POSIXt} objects are converted to annual values by multiplying slope
estimates by multiplying by 31556926 seconds / year (with a warning
about rounding error). No scaling is done for \code{.mode == 'time'} or
\code{.mode == 'count'}.}

\item{t_fit}{Should the underlying generalized linear model's correlation
structure be fit based on the time coordinate (.dt), or only on the
sequence of observations in the data? Setting this to TRUE is safer if
you are uncertain of the sequence of observations in the source data,
or significant missing values in the data, but it significantly slows
model fitting.  Results tend to be very similar for complete or near
complete data.}

\item{retain_model}{Boolean.  If \code{TRUE}, the piecewise linear model is
returned via the 'model' slot of the return value.}
}
\value{
a list, with the following components:
\describe{
\item{Sample}{Sample size on which model is based}
\item{L.Ratio}{likelihood ratio comparing the piecewise linear model to a
simple linear model.}
\item{p-value}{the p-value (by ANOVA) comparing the two models.}
\item{delta-AIC}{Change in AIC between the two models.  A value less than
about -2 is evidence that the piecewise model would outperform the
strict linear model for "predicting" past sea levels.}
\item{slope_old}{The estimated slope of changes in sea level prior to the
"recent" period specified via the .span parameter.}
\item{slope_old_err}{Standard error of that slope.}
\item{slope_recent}{Estimated slope during the "Recent" period,  This is
calculated as the sum of two model parameters, the overall slope and
a "recent" correction.}
\item{slope_recent_err}{Standard error of the recent slope.  This is
calculated using the formula for the variance for the sum of two
uncorrelated random variables, treating the overall slope and
recent slope correction as uncorrelated.}
\item{model}{The underlying piecewise GLS model, returned only if
retain_model == TRUE}
}
}
\description{
Fit a piecewise linear model to test whether a recent change in
rate of sea level rise has occurred.
}
\details{
A frequent question raised in analysis of recent sea level records is whether
the rate of sea level rise is increasing, as predicted by numerous climate
models over the years. This is a simple question that is rather more complex
to answer statistically than it at first appears.

This function checks for a change is rate of sea level rise by fitting a
piecewise linear model and testing if a change in slope at a (user-defined)
breakpoint improves the statistical fit enough to matter.  A piecewise linear
model fits the data with (here) two linear segments that join at a breakpoint
or knot. The underlying model is a generalized linear model, with an AR1
autocorrelated error term.

The AR1 process, an "autoregressive process of order 1", is a simple
assumption about correlation structures that reduces the risk of overstating
the precision of trend estimates. Actual tidal data series tend to have more
complex autocorrelation structure. If working with high frequency data,
short-term structure should be removed before using this function, typically
by calculating average sea level over convenient and meaningful time periods,
like months or years.  See additional comments on use of the AR1 model in
analysis of sea level data in the help file for \code{slr_slope()}.

This function is ripe for abuse. The problem is that rates of sea level
rise year over year vary.  By chance, some periods of the historic record
will have higher rates of sea level rise than others.  This makes it far too
easy to "cherry pick" starting dates for a recent period of record, and claim
a statistically meaningful change in rate of sea level rise.  Use with
appropriate humility.

For a more robust approach (with lower statistical power, but deeper
justification), see the function \code{all_sr_change()}
}
\examples{

}
\seealso{
Other sea level rate functions: 
\code{\link{all_slr_change}()},
\code{\link{get_sl_trend}()},
\code{\link{slr_slope}()}
}
\concept{sea level rate functions}
