% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLRSIM-package.R
\docType{package}
\name{SLRSIM-package}
\alias{SLRSIM}
\alias{SLRSIM-package}
\title{SLRSIM: Tools to analyse sea level rise}
\description{
SLRSIM offers tools to facilitate analysis of recent sea level
             rise trends and forecasting of future flood frequency under 
             user-specified sea level rise scenarios. The package includes 
             tools to access NOAA tide guage data, examine long-term trends
             in sea leve rise, compare recent sea level rise to previous periods
             in the data record, and forecast future flood frequency based on 
             historical data.
}
\details{
You can add additional information about the package here.
}
\author{
\strong{Maintainer}: Curtis Bohlen \email{curtis.bohlen@maine.edu}

}
\keyword{internal}
