% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_slr_change.R
\name{all_slr_change}
\alias{all_slr_change}
\title{Place recent rate of sea level change into historical context}
\usage{
all_slr_change(
  .data,
  .sl,
  .dt,
  .span = 20L,
  .interval = 1L,
  .mode = c("year", "duration", "count"),
  t_fit = FALSE
)
}
\arguments{
\item{.data}{Source data frame for data.  Use NULL if no data frame is used
and all data is passed from the enclosing environment.}

\item{.sl}{The data variable (usually found in the data frame) showing sea
level or mean sea level.  Must be a named variable, not an
expression.}

\item{.dt}{Data variable containing corresponding midpoint dates for the
period of averaging used to calculate .sl. Must be a named variable,
not an expression. Midpoint dates for a given month can be
approximated with \verb{as.Date(paste(year, month, 15, sep = '-')}.}

\item{.interval}{Integer.  Default = 1.  Spacing of samples on which to
calculate slopes.}

\item{.mode}{one of c('year', 'duration', 'count') indicating whether the results should be scaled to
annual values by multiplying by 365.25. If \code{.dt} is not a Date, this
is ignored, and no scaling is conducted.}

\item{t_fit}{Should models be fit based on time coordinates, or
only on the sequence of observations in the data?  Setting this
to TRUE is safer if you are uncertain of the sequence of observations
in the source data, or if there are many missing values in your data,
but it slows model fitting somewhat.}

\item{.span.}{Span defining the "recent" period for fitting the model.
integer or difftime object.  Interpretation of .span depends on the
value of the by_year. If \code{by_year == TRUE}, \code{.span} must be an
integer, and is interpreted as the number of years included in the
"recent" period for fitting the model.  If \code{by_year == FALSE}, and
\code{.span} is an integer, it is interpreted as the number of records to
include in the 'recent' period.}
}
\value{

}
\description{
Compares the recent rate of sea level rise to rates for all prior periods of
the same length.
}
\details{
A frequent question raised in analysis of recent sea level records is whether
the rate of sea level rise is increasing, as predicted by numerous climate
models over the years. This is a simple question that is rather more complex
to answer statistically than it at first appears.

This function checks for a change in rate of sea level rise in the most
recent period of .span years by comparing it to every prior period of the
same length in the historic record.

This function returns a probability that corresponds to the percentage of
all prior periods with a slope equal to or greater than the most recent
slope.

While conceptually, this is akin to a p-value, with the null hypothesis that
the current period is drawn from the same distribution as all prior slopes,
it should not be interpreted strictly. Consecutive slopes are strongly
auto-correlated, and this analysis makes no effort to account for that.
}
\examples{

}
\seealso{
Other rate changes: 
\code{\link{slr_change}()}
}
\concept{rate changes}
