% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epoch.R
\name{get_epoch}
\alias{get_epoch}
\title{Retrieve the tidal epoch}
\usage{
get_epoch(.station)
}
\arguments{
\item{.station}{}
}
\value{
A named list of two integers, representing the starting and ending
years of the official tidal epoch.
}
\description{
For most US tidal stations, the function currently returns
\code{c(start = 1983, end = 2001)} but some stations on the Gulf Coast and Alaska
have "anomalously high" sea level rise rates, and so tidal forecasts have
been calibrated based on a more recent five year period, usually 2012 through
2016.  For those Stations, this function will return
\code{c(start = 2012, end = 2016)}.
}
\details{
The function is especially useful for international stations, where the
period used to calculate tidal statistics my be different.  NOAA is expected
to adopt a revised tidal datum in 2025.
}
\examples{
Retrieve HAT for Portland, Maine
portland_id <- 8418150
get_epoch(portland_id)   # c(start = 1983, end = 2001)
}
\seealso{
Other station information access functions: 
\code{\link{get_datums}()},
\code{\link{get_hat}()},
\code{\link{get_sl_trend}()}
}
\concept{station information access functions}
