% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_slope.R
\name{slr_slope}
\alias{slr_slope}
\title{Calculate Long Term Estimate of SLR}
\usage{
slr_slope(.data, .msl, .dt)
}
\arguments{
\item{.data}{Source data frame}

\item{.msl}{Data variable showing sea level or mean sea level}

\item{.dt}{Data variable containing corresponding midpoint dates of the
period of averaging for .msl.}
}
\value{

}
\description{
Calculate Long Term Estimate of SLR
}
