% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLRSIM-package.R
\docType{package}
\name{SLRSIM-package}
\alias{SLRSIM}
\alias{SLRSIM-package}
\title{SLRSIM: Tools to Analyse Sea Level Rise}
\description{
SLRSIM offers tools to facilitate analysis of recent sea level
    rise trends and forecasting of future flood frequency under user-specified 
    sea level rise scenarios. The package includes tools to access NOAA tide
    gauge data, examine long-term trends in sea level rise, compare recent sea 
    level rise to previous periods in the data record, and forecast future flood
    frequency based on historical data.
}
\details{
The Package originated in code developed for the 2020/2021 State of Casco
Bay Report. In response to requests from colleagues, I first extended
our analysis to a couple of other NOAA tide stations, and then developed this
package to capture key ideas and code for reuse.

The intent was to offer a collection of tools to facilitate rapid preparation
of historical analyses and forecasts principally to facilitate communication
about sea level rise to policy makers and the public.

Since I first began working on the package, NOAA has begun to release
related analyses for many NOAA stations, so the need for the package is less
pressing (when workiong with NOAA data), but this package offers the user
fine-grained control not available from the prepared analyses offered by
NOAA.
}
\author{
\strong{Maintainer}: Curtis Bohlen \email{curtis.bohlen@maine.edu}

}
\keyword{internal}
